Factor conversions of results into two base classes.
We are going to introduce another variant of the Expr class. By moving
the conversion operators to a base class, we can avoid code duplication.

Furthermore, the plan is to turn class Result into a template. Then we
want to have all implementations factored into a non-templated base class,
ResultBase. This prepares for such an implementation.