Evaluate Tcl commands in class Expr's destructor and permit exceptions.
So far the commands were executed from class Command's destructor. The
problem is that destructors are noexcept(true) by default since C++11.
When a Tcl error occurs, the exception that is thrown from the evaluator
terminates the program via terminate(), but does not escape from the
destructor.

Move the evaluation to class Expr and mark its destructor noexcept(false).
Just marking Command's destructor as noexcept(false) is not sufficient
because objects held in a shared_ptr must not throw exceptions from the
destructor (see http://www.eel.is/c++draft/util.smartptr.shared.const).
Therefore, the only option is to move the evaluation.