Store callbacks in std::function<>.
The current implementation of callbacks requires to specify a function
pointer. But it would be much more convenient if arbitrary callables could
be specified.

Rewrite the implementation to use std::function<>. This permits lambda
functors to be used as callbacks.

As before, the cases where a functor is passed as callback, the
implementation is a function template. In some cases, this template is one
of several overloads. Then one overload typically takes a
const std::string& in place of the functor.

Now consider the case where a non-const std::string variable is the
callback argument:

   std::string command = callback(...);
   after(delay, command);

In this case, the function template, whose parameter is Functor&& is a
better fit than the other overload that takes a const std::string&. This
causes the constructor of the Callback object to fail because it cannot
convert the std::string to a std::function<>.

The solution is to take the function template out of the overload set when
the argument cannot be converted to std::function<>. This is the task of
the RequireFunctor<> helper function. It uses SFINAE via std::enable_if<>.