Permit expressions bound with 'bind' to return a completion code.
The implementation checks whether the return value of the functor is the
same as the expected return value of the functor type. To do this check,
we have to deduce the return type of the actual functor. But this works
only for function objects, not when the functor expression is a (Tcl)
string. This requires us to use if constexpr before the is_same_v<> check.
And that pushes the compiler requirement to C++17.